产地、分类向右移了点; 品牌分类过滤； 产品页logo居中； seo title

+++ src/com/knx/controller/brand/CategoryController.java
@@ -33,7 +33,7 @@ public class CategoryController {
 				for(CategoryDTO dto: categoryDTOList){
 					dto.setShow(false);
 					for(CategoryDTO showDto: dtoList){
-						if(dto.getIdCat()==showDto.getIdCat()){
+						if(dto.getIdCat().longValue()==showDto.getIdCat().longValue()){
 							dto.setShow(true);
 						}
 					}
